requirements.txt
argcomplete==3.4.0
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.11.1
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
commitizen==3.28.0
cssselect2==0.7.0
decli==0.6.2
decouple==0.0.7
Django==5.0.7
django-classy-tags==4.1.0
django-cms==4.1.2
django-entangled==0.5.4
django-filer==3.1.4
django-formtools==2.5.1
django-fsm==2.8.2
django-parler==2.3
django-polymorphic==3.1.0
django-sekizai==4.1.0
django-treebeard==4.7.1
djangocms-admin-style==3.3.1
djangocms-alias==2.0.1
djangocms-attributes-field==3.0.0
djangocms-file==3.0.1
djangocms-frontend==1.3.3
djangocms-googlemap==2.2.0
djangocms-picture==4.1.1
djangocms-snippet==3.1.1
djangocms-style==3.1.0
djangocms-text-ckeditor==5.1.5
djangocms-versioning==2.0.2
djangocms-video==3.1.0
easy-thumbnails==2.9
gunicorn==23.0.0
h11==0.14.0
html5lib==1.1
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
lxml==5.2.2
MarkupSafe==2.1.5
numpy==2.1.1
opencv-python==4.10.0.84
outcome==1.3.0.post0
packaging==24.1
pillow==10.4.0
pinscrape==3.2.3
pluggy==1.5.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.9
pydotmap==0.1.3
PySocks==1.7.1
pytest==7.2.0
python-decouple==3.8
python-dotenv==1.0.1
pyunsplash==1.0.0rc2
PyYAML==6.0.1
questionary==2.0.1
reportlab==4.2.2
requests==2.32.0
selenium==4.24.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.1
svglib==1.5.1
termcolor==2.4.0
tinycss2==1.3.0
tomlkit==0.13.0
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
wsproto==1.2.0
